ci(workflows): add trigger for Docker Publish Workflow on new release creation ci(workflows): update Docker image build and push actions to include version tag ci(workflows): add support for building and pushing Docker image with nightly tag
The changes were made to enhance the CI workflows related to creating new releases and publishing Docker images. The trigger for the Docker Publish Workflow was added to the create-new-release.yml file to automate the process of publishing Docker images when a new release is created. Additionally, the publish-docker-image.yml file was updated to include support for building and pushing Docker images with version tags and a nightly tag based on the event payload. This improves the automation and versioning of Docker images in the CI pipeline.